+++ src/Client.cpp
@@ -85,6 +85,13 @@ void Client::registerCapability(const std::string& registrationId, const std::st
     sendRequest(nextRequestId++, "client/registerCapability", lsp::RegistrationParams{{registration}});
 }
 
+void Client::unregisterCapability(const std::string& registrationId, const std::string& method)
+{
+    lsp::Unregistration unregistration{registrationId, method};
+    // TODO: handle responses?
+    sendRequest(nextRequestId++, "client/unregisterCapability", lsp::UnregistrationParams{{unregistration}});
+}
+
 ClientConfiguration Client::getConfiguration(const lsp::DocumentUri& uri)
 {
     auto key = uri.toString();